@@ -51,8 +51,8 @@
 		<artifactId>maven-compiler-plugin</artifactId>
 		<version>3.5.1</version>
 		<configuration>
-			<source>1.7</source>
-			<target>1.7</target>
+			<source>8</source>
+			<target>8</target>
 		</configuration>
 	</plugin>
 	</plugins>
