@@ -2,25 +2,26 @@ package start;
 
 import food.Food;
 import org.joda.time.DateTime;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 import storage.Shop;
 import storage.Trash;
 import storage.Warehouse;
-import static org.junit.Assert.assertThat;
+
+import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
 
 /**
  * Unit test for ControllQuiality.java.
- * It test moving food in trash, shop and warehouse.
+ * It is testing moving food in trash, shop and warehouse.
  */
-public class ControllQualityTest {
+class ControllQualityTest {
 
 
     /**
-     * When try move good fruit to warehouse should check than warehouse save it.
+     * When try moving good fruit to warehouse should check than warehouse save it.
      */
     @Test
-    public void whenTryMoveGoodFruitToWarehouseShouldCheckThatAtWarehouse() {
+    void whenTryMoveGoodFruitToWarehouseShouldCheckThatAtWarehouse() {
         ControllQuality control = new ControllQuality();
         DateTime expairTime = new DateTime();
         expairTime = expairTime.plusMonths(3);
@@ -38,7 +39,7 @@ public class ControllQualityTest {
      * When try move some food to shop should check than shop is save it.
      */
     @Test
-    public void whenTryMoveSomeFoodToShopShouldCheckThatFoodOnTheShop() {
+    void whenTryMoveSomeFoodToShopShouldCheckThatFoodOnTheShop() {
 
         ControllQuality control = new ControllQuality();
         DateTime createTime = new DateTime();
@@ -56,10 +57,10 @@ public class ControllQualityTest {
     }
 
     /**
-     * When try move some food which must with discount should check that shop save it.
+     * When try moving some food which must with discount should check that shop save it.
      */
     @Test
-    public void whenTryMoveSomeFoodWhichMustWithDiscountShouldCheckThatIsCorrect() {
+    void whenTryMoveSomeFoodWhichMustWithDiscountShouldCheckThatIsCorrect() {
         ControllQuality control = new ControllQuality();
         DateTime createTime = new DateTime();
         createTime = createTime.plusDays(4);
@@ -76,10 +77,10 @@ public class ControllQualityTest {
     }
 
     /**
-     * When try move some food which must be at the trash should check that controller move it to trash.
+     * When try moving some food which must be at the trash should check that controller move it to trash.
      */
     @Test
-    public void whenTryMoveSomeFoodWhichMustAtTheTrashShouldCheckThatTrashHaveIt() {
+    void whenTryMoveSomeFoodWhichMustAtTheTrashShouldCheckThatTrashHaveIt() {
         ControllQuality control = new ControllQuality();
         DateTime expaireTime = new DateTime();
         expaireTime = expaireTime.plusDays(1).plusHours(23);
